Refactor: Lotto필드 수정 및 입력 오류 처리 컨트롤러로 이동 Lotto 의 필드 Set -> TreeSet으로 변경 InputHandler 삭제 후 컨트롤러에서 입력받고 오류 처리를 하도록 변경
